requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
python-jose
passlib[bcrypt]
alembic
pydantic
psycopg2-binary
